WTF::Vector: Fix primitive types identified as not safe to memcpy.
The culprit was the implementation of IsAssignable<> traits.

Specifically, "std::declval<T>() = X" is invalid if T is a primitive
type, because the assignment of primitives is not an overloaded
operator and the compilability of this statement for a primitive T is
different from that for a non-primitive T.

The fix is to simply change the left-hand side of the assignment to
"std::declval<T&>".

BUG=624642

Review-Url: https://codereview.chromium.org/2110183005
Cr-Original-Commit-Position: refs/heads/master@{#403131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 757e4b7379983a8760fe8f91e40bb3b34d64bde4