PlzNavigate: update test filters
This CL updates the test filters for the FYI bot LinuxBrowserSideNavigation.
This follows the change for ServiceWorker support where we removed our first
implementation that was only partially working.
Making this NOTRY since there are no code changes and this is for an FYI bot.

BUG=504347, 576261
TBR=carlosk@chromium.org
NOTRY=TRUE

Review-Url: https://codereview.chromium.org/2110263003
Cr-Original-Commit-Position: refs/heads/master@{#403171}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb17a07a7aa2cf0bf6fe29c9f65117a2d726cd90