Reinstate BoxPainter's unrederable background rrect logic
http://crrev.com/2008913002 replaced the unrenderable case with an
assert, on the assumption that higher-level logic should ensure a
renderable rrect.

But in certain cases we compute the background rrect by shrinking the
border rrect - and this can yield degenerate/unrenderable results when
some radii are not large enough to accomodate the requested inset.

E.g.

 width: 100
 left radius: 0
 right radius: 100
 border-width: 10

We attempt to inset and shrink all radii by 5 (half border-width), but
we can only reduce the right radius since the left one is already zero.
So we end up with

 width: 90
 left radius: 0
 right radius: 95

which is not renderable.

This CL restores the clip-based fallback for unrenderable rrects.

BUG=614825
R=chrishtr@chromium.org

Review-Url: https://codereview.chromium.org/2112643002
Cr-Original-Commit-Position: refs/heads/master@{#403321}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 804f289c635c1ccf84e8079d4a5c78d4d8786516