Fix date picker datalist suggestion dropdown size with use-zoom-for-dsf.
When a date picker has a suggestion list it's supposed to be the same
width as the input element. With use-zoom-for-dsf on I discovered that
it was too wide because the width that was being passed in was using
"root frame" coords that were not scaled correctly.

BUG=621221

Review-Url: https://codereview.chromium.org/2095113002
Cr-Original-Commit-Position: refs/heads/master@{#403394}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e94f26e8f22e96fda6ade9ccb35380906ee77ab