Fix rounding of phase for background image
In the BackgroundImageGeometry::calculate methods we were previously
computing the full image tiling phase value and then rounding to int
at the end. But frequently the phase is computed as the negative of a
background-position value, which is itself frequently negative. The
result of negating a value then rounding is not the same as rounding
then negating, and this difference matters when trying to, for
instance, match the end of a linear gradient with a solid background
color.

This patch changes phase rounding to round before negating. The test
case is fixed, and no other test results change.

R=pdr@chromium.org,fmalita@chromium.org
BUG=622294

Review-Url: https://codereview.chromium.org/2103563002
Cr-Original-Commit-Position: refs/heads/master@{#403458}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 689f41251eab6e3812c09296a14cb1edc6db3b0e