Change code path for structured cloning ImageBitmap
This is how it works currently:
When an ImageBitmap is un-premultiplied, we read bitmap's pixel data in
premultiplied format. When de-serialize happens, we call
ImageBitmap::create(ImageData*) and in that create() function, we create
an SkImage that is in premultiplied format because the input data is
already premultiplied. This is causing the problem because the SkImage
should be in un-premultiplied format.

We change the implementation like this:
When we serialize the ImageBitmap into uint8_t*, we read the pixel
data according to whether the ImageBitmap is premultiplied or not,
also we store the information whether the bitmap is premultiplied or
not. When de-serialization happens, we design a new ImageBitmap
constructor specifically for it.

Notice that after this change, the layout tests
fast/canvas/webgl/texImage-imageBitmap-structured-clone.html
passes.

BUG=615172

Review-Url: https://codereview.chromium.org/2039983002
Cr-Original-Commit-Position: refs/heads/master@{#403475}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58c08f3acefbd7a44387d03237967a4644130c0f