Activate Complex text on Android after memory improvements
In the blocking bugs of crbug.com/617568 work was done on improving font
memory managament: Fixing the ref counting of the FontDataCache, zero
copy access to font blobs and purging the font cache on low memory
pressure. So it is time to give it another try and see whether this
addresses the memory regressions observed on bots when activating
complex text on Android seen in crbug.com/577306.

TBR'ing to get bot results. We will revert again if the regression still
occurs.

BUG=561099
TBR=eae

Review-Url: https://codereview.chromium.org/2120143002
Cr-Original-Commit-Position: refs/heads/master@{#403648}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6867727d32a96418c1c34128251807ff712aff81