Make sure RecordingImageBufferSurface's current frame is valid
The "FallbackReasonUnknown" should not happen. Right now it could still
occur in RecordingImageBufferSurface::finalizeFrameInternal when the
m_currentFrame is null (theoretically).
In this CL, we change ASSERT to CHECK, to make sure that the
m_currentFrame is never null in the finalizeFrameInternal() method.

BUG=585578

Review-Url: https://codereview.chromium.org/2115493004
Cr-Original-Commit-Position: refs/heads/master@{#403707}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9c8483972fea54016b0cf05e5e50fb8cba47248d