Always use the WebFrameWidget when attaching the root graphics layer or the compositor animation timeline in ChromeClientImpl.
Now that we always instantiate a WebFrameWidget for local
roots, we can remove the fallback code that goes directly
through WebView.

This change also fixes an issue where local roots were not
being passed to
ChromeClientImpl::attachCompositorAnimationTimeline which
caused the AnimationTimeline to be attached to the wrong
widget.

BUG=587025

Review-Url: https://codereview.chromium.org/2036403002
Cr-Original-Commit-Position: refs/heads/master@{#403781}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af41ee000f480c570adf535fede679c90819d10f