Prevent to measure the whole word when break-all
This patch applies the same heuristic fix as break-word to break-all.

Since the switch to the complex path, r385693 fixed the performance and
correctness issue of break-word and break-all in common cases. However,
it also introduced a performance regression when a word is really long,
such as minimized JS.

This can be fixed by taking over the ShapeResult to the next lines as
Gecko does, but it will require larger overhaul of the line breaker and
need more time to bake.

As an interim fix, r387974 introduced a heuristic fix for break-word.
This patch applies the same heuristic to break-all.

This patch also revealed that the fix for crbug.com/603679 was not
correct. MidWordBreak must be recomputed when floats pushed the line
down and available width changed. This patch includes the fix.

BUG=622810, 603679

Review-Url: https://codereview.chromium.org/2077313002
Cr-Original-Commit-Position: refs/heads/master@{#403830}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d24d249700deef0ed63bf12ebb6af3b28f8225fb