Moved RootFrameViewport creation to happen after document attachment
I previously placed this to happen just before the first layout but that was
overly conservative; we really just need the document to be attached. This will
allow constructing a ViewportScrollCallback for a FrameView in just one place
for both root and nested frames. This will be needed in the setRootScroller
proposal in future patches.

BUG=505516

Review-Url: https://codereview.chromium.org/2123663003
Cr-Original-Commit-Position: refs/heads/master@{#403973}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3efe8736a302182bff2e9101044fd8e15668bbb6