Teach HistoryItem about a request's initiator.
Currently, we don't store enough information on HistoryItem to deal
correctly with 'SameSite' cookies for cases in which the user navigates
their history. This patch adds the initiator of the original request to
the HistoryItem, and restores it as part of the navigation work in
FrameLoader.

BUG=626243
R=yoav@yoav.ws

Review-Url: https://codereview.chromium.org/2130623003
Cr-Original-Commit-Position: refs/heads/master@{#404123}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c19960b67391c5789878f475163b813ad9daf994