Moved viewport layer registration into WebViewImpl.
This is currently done inside VisualViewport by passing it a pointer to the
WebLayerViewView and letting it register all the viewport related layers.

This isn't the right place to do it since not all the layers are related to the
VisualViewport. Additionally, it requires VisualViewport to be more aware of
compositing than it should be.

BUG=505516

Review-Url: https://codereview.chromium.org/2128363002
Cr-Original-Commit-Position: refs/heads/master@{#404359}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2c06cf000f191caa07b13be0537844cf6448402