Create the harness and shell of update_test_expectations
This patch creates the main classes and stubs out the main methods. The meat of the
process is left for the next patch. update_test_expectations is a script that looks
at the TestExpectations file and checks each line against results from the build
waterfall to determine if a line is still flaky. Non-flaky entries in TestExpectations
are then removed.

BUG=595414

Review-Url: https://codereview.chromium.org/2119303003
Cr-Original-Commit-Position: refs/heads/master@{#404402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d70d9e2e8f1c9efb80d1dd1146816d5f265f367a