Ensure SVG PaintOffset is reset at the HTML/SVG boundary [spv2]
This patch fixes a crash on https://codepen.io where the paint offset
in PaintPropertyTreeBuilder::updateSvgLocalToBorderBoxTransform would
not be reset if transformToPixelSnappedBorderBox was identity.

The new testcase, SvgPixelSnappingShouldResetPaintOffset, both ensures
the DCHECK in PaintPropertyTreeBuilder::updateTransform is not hit, and
verifies the transform property tree is correct in this case.

BUG=600618

Review-Url: https://codereview.chromium.org/2131803002
Cr-Original-Commit-Position: refs/heads/master@{#404483}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 306affbd20fa8e16b8dd7677061fabde355ead8d