Correct color bleeding of no-repeat pattern in canvas2d
Currently the image pattern shader created is only padded
with 1 black transparent pixel right and bottom. We should
pad it on the left and top as well so that when pattern is
transformed, the color on its left and top will not bleed.

BUG=602509

Review-Url: https://codereview.chromium.org/2130063002
Cr-Original-Commit-Position: refs/heads/master@{#404491}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c2635f87e413ac3f81a81ac1e37fce54fdc8d26d