[Mac] Draw the caps lock indicator on the left side for RTL direction.
For RTL direction, the caps lock indicator for password fields should be drawn
on the left side of it, but currently it is always drawn on the right side. This
CL makes it possible to draw the caps lock indicator on the left or right side
depend on current writing direction.

BUG=623784

Review-Url: https://codereview.chromium.org/2125533002
Cr-Original-Commit-Position: refs/heads/master@{#404665}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dec7875f088d45cd79a82a985c14dfdc7b2c1825