Remove LocalFrameLifecycleObserver::willDetachFrameHost()
willDetachFrameHost() is dispatched at almost the same timing as contextDestroyed().
Hence this CL replaces willDetachFrameHost() with contextDestroyed().

The only exception is LocalDOMWindow::willDetachFrameHost().
It needs to be called before the frame is actually detached from the frame host,
so this CL moves the contents to LocalFrame::detach().

BUG=610176

Review-Url: https://codereview.chromium.org/2142753002
Cr-Original-Commit-Position: refs/heads/master@{#404924}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1d739c3a4cd78b3a4aa5a8a6ebf313df2dca1e3