Add heuristic to LayoutTextControl::hasValidAvgCharWidth()
This patch adds a heuristic check to see if the font has invalid
avgCharWidth.

The spec is silent about how avgCharWidth is computed for CJK fonts,
and some fonts selected to match to CJK full-width characters.

Since in most fonts, avgCharWidth is close to the width of "0", when
it's too big compared to the width of "0" can be considered as invalid.

BUG=625058

Review-Url: https://codereview.chromium.org/2145903002
Cr-Original-Commit-Position: refs/heads/master@{#405086}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77183f3386fcb2362ca6140ab5fb15958d7cfa2d