Rewrite test in terms of promise rejection
This patch rewrites animation-finish-event-cancelled.html to listen for
rejection of the animation's finished promise instead of using a combination
of onfinish and assertions about code being unreached. This makes the test
easier to read and brings it more in line with the style of tests in the
web-platform-tests suite.

BUG=623437

Review-Url: https://codereview.chromium.org/2145263002
Cr-Original-Commit-Position: refs/heads/master@{#405384}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d8a910760e0c33dc93b4751e330b729936a1df5c